+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [v2.2.0] - 2026-03-20
+
+- Add "Collapse Pinned Comments" setting to auto-collapse pinned/stickied comments
+- Add "Can't sign in?" troubleshooting section
+- Liquid Glass: fix blur overlay issue when viewing messages
+
 ## [v2.1.0] - 2025-03-12
 
 - Add Steam store deep linking support (thanks @wdeezy for the contribution!)
@@ -272,6 +278,7 @@ There are currently a few limitations:
 ## [v1.0.0] - 2023-10-13
 - Initial release
 
+[v2.2.0]: https://github.com/JeffreyCA/Apollo-ImprovedCustomApi/compare/v2.1.0...v2.2.0
 [v2.1.0]: https://github.com/JeffreyCA/Apollo-ImprovedCustomApi/compare/v2.0.0...v2.1.0
 [v2.0.0]: https://github.com/JeffreyCA/Apollo-ImprovedCustomApi/compare/v1.4.5...v2.0.0
 [v1.4.5]: https://github.com/JeffreyCA/Apollo-ImprovedCustomApi/compare/v1.4.4...v1.4.5